// src/app/enums/unit-of-measure.enum.ts
export enum UnitOfMeasureType {
  SACA = 'SACA',
  FARDO = 'FARDO',
  UNIDADE = 'UNIDADE', 
  KG = 'KG',
  GRAMAS = 'GRAMAS',
  METROS = 'METROS',
  CENTIMETROS = 'CENTIMETROS',
  LITROS = 'LITROS',
  MILILITROS = 'MILILITROS'
}

export enum UnitOfMeasureAbbreviation {
  S = 'S',
  F = 'F',
  UN = 'UN',
  KG = 'KG',
  G = 'G',
  M = 'M',
  CM = 'CM',
  L = 'L',
  ML = 'ML'
}